CSS: narrow body text
WCAG AAA guidelines encourage limiting text to 80 chars. Unlike A and
AA, the AAA level is more of a list of suggestions than a requirement.
Most other studies seem to indicate 70 is a good minimum but 100 is a
bit excessive.

"ch" units are broken on NetSurf, so I went with the closest "em"
approximation (since I already use "em" everywhere else).