Remove "last deployed" date
Visitors can see that on post update dates; I can see that in
builds.sr.ht.

This results in not having to upload every page on a trivial change. Add
the "-c" rsync flag to take advantage of that. Doing so should also
result in longer-lived cached copies, with etags and last-modified
headers.