Reduce accidental skips, add transcript, microdata
When a figure is in the middle of a section, mobile screen reader users
might not switch away from heading-based navigation. Instead, they might
just jump to the next heading.

Re-organize the page to put some skippable content closer to the end of
sections, and add a section describing the need for this.

Also add a transcript to the screenshot of BMFW, and adopt the new
quotecaption shortcode, and fix some structured data by using some
shortcodes properly.