Use Flatpak Chromium headless
inb4 a bunch of people from the privsec community @ me for using flatpak
chromium:

Yes, the sandboxing is different, but I'm not using this as my
day-to-day browser. I'm using it headless for measurements and auditing.
I want a browser that behaves as close to upstream official Chromium
builds when doing Chromium measurements.

Take some deep breaths. Here, have a glass of water.